requirements.txt
pandas
rdkit-pypi
biopython
pubchempy
